Move creatable playbook types to /models
