+++ src/functions/playbooks/create-one-playbook.ts
@@ -6,16 +6,14 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { encode as base64Encode } from 'base-64';
-import { RawPlaybookDecodedMetadata } from '../../models';
-import { isUUID, Markdown, NumericID, RawNumericID, toRawNumericID, UUID } from '../../value-objects';
+import { CreatablePlaybook, toRawCreatablePlaybook } from '../../models';
+import { UUID } from '../../value-objects';
 import {
 	APIFunctionMakerOptions,
 	buildHTTPRequest,
 	buildURL,
 	fetch,
 	HTTPRequestOptions,
-	omitUndefinedShallow,
 	parseJSONResponse,
 } from '../utils';
 
@@ -39,52 +37,4 @@ export const makeCreateOnePlaybook = (makerOptions: APIFunctionMakerOptions) => 
 			throw Error('Unknown error');
 		}
 	};
-};
-
-export interface CreatablePlaybook {
-	userID?: NumericID;
-	groupIDs?: Array<NumericID>;
-
-	name: string;
-	description?: string | null;
-	labels?: Array<string>;
-
-	isGlobal?: boolean;
-
-	body: Markdown;
-	coverImageFileGUID?: UUID | null;
-}
-
-interface RawCreatablePlaybook {
-	Name: string;
-	Desc: string | null;
-	Body: string; // Base64 encoded markdown string
-	Metadata: string; // Base64 encoded RawPlaybookDecodedMetadata
-
-	// !WARNING: That's not working right now, CHECK
-	UID?: RawNumericID;
-	GIDs: Array<RawNumericID>;
-
-	Global: boolean;
-	Labels: Array<string>;
-}
-
-const toRawCreatablePlaybook = (creatable: CreatablePlaybook): RawCreatablePlaybook => {
-	const metadata: RawPlaybookDecodedMetadata = { dashboards: [] };
-	if (isUUID(creatable.coverImageFileGUID))
-		metadata.attachments = [{ context: 'cover', type: 'image', fileGUID: creatable.coverImageFileGUID }];
-
-	return omitUndefinedShallow({
-		UID: creatable.userID ? toRawNumericID(creatable.userID) : undefined,
-		GIDs: creatable.groupIDs?.map(id => toRawNumericID(id)) ?? [],
-
-		Global: creatable.isGlobal ?? false,
-		Labels: creatable.labels ?? [],
-
-		Name: creatable.name,
-		Desc: creatable.description ?? null,
-
-		Body: base64Encode(creatable.body),
-		Metadata: base64Encode(JSON.stringify(metadata)),
-	});
 };
